refactor: highlightText 유틸함수를 shared/lib 폴더로 이동
